requirements.txt
praw
vaderSentiment
alpaca-trade-api
schedule
python-dotenv
pytest
requests